ovs: add default 30s timeout to ovs-vsctl operations
If the vswitchd exits (or is OOM killed) while an ovs-vsctl is
being run, that command may hang waiting for the database state
to be updated, which never happens because vswitchd is dead and
hasn't been restarted.  Add a timeout so we don't block forever.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1571379